riscv: kexec_file: Split the loading of kernel and others
mainline inclusion
from mainline-v6.16-rc1
commit f3cc9d65c88734253e0f7bfaf8b96583089b424b
category: bugfix
bugzilla: https://github.com/RVCK-Project/rvck/issues/70

--------------------------------

This is the preparative patch for kexec_file_load Image support.

It separates the elf_kexec_load() as two parts:
- the first part loads the vmlinux (or Image)
- the second part loads other segments (e.g. initrd,fdt,purgatory)

And the second part is exported as the load_extra_segments() function
which would be used in both kexec-elf.c and kexec-image.c.

No functional change intended.

Link: https://lore.kernel.org/r/20250409193004.643839-2-bjorn@kernel.org
